requirements.txt
asgiref==3.8.1
convertdate==2.4.0
Django==5.1.3
djangorestframework==3.15.2
jalali_core==1.0.0
jdatetime==5.0.0
psycopg==3.1.8
psycopg-binary==3.1.4
psycopg-pool==3.1.2
psycopg2==2.9.9
PyMeeus==0.5.12
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
